Add generateDocument function to main.js. Pass commonData as a initial value for components
