--- src/components/SaveButton/SaveButton.scss
@@ -1,3 +1,0 @@
-.SaveButton {
-
-}
